setup moved out of pipeline
